Suppress type comparison warning in RichEnum.lookup.
A reasonably common pattern is a RichEnumValue that has an attriibute that's
a list of some other RichEnumValue, usually related values. lookup() can then
be used to search said attribute for a specific member.

We always check up front to see if the object we're searching for and the
attribute value in question are exactly matching; in this case that generates
a type comparison warning that's spurious, since the next case (checking if
the attribute value is an iterable and seeing if the object we're searching
for is in it) does the correct thing.